Add files via upload
Predicting the life condition of a component is so crucial for industries that have intent to grow in a fast paced technological environment. Recent studies on predictive maintenance help industries to create an alert before the components are corrupted. To predict and be ahead of all these scenarios this project implies Artificial Intelligence (AI) and Deep Learning using long short-term memory(LSTM) neural networks to predict when an aircraft engine might require to be serviced or replaced. 
For these predictions we require a dataset of previous aircraft historical data having 21 sensor values of each aircraft. Aircrafts have three different settings = set1, set2, set3 so that we can manipulate the data and find out the relation/ trend in the data so that our system gets capable of predicting remaining useful life (RUL) of an aircraft engine.